+++ vcd/provider.go
@@ -128,7 +128,8 @@ func Provider() terraform.ResourceProvider {
 			"vcd_nsxv_dnat":          resourceVcdNsxvDnat(),         // 2.5
 			"vcd_nsxv_snat":          resourceVcdNsxvSnat(),         // 2.5
 			"vcd_nsxv_firewall_rule": resourceVcdNsxvFirewallRule(), // 2.5
-			"vcd_ipset":              resourceVcdIpSet(),            // 2.6
+			"vcd_nsxv_dhcp_relay":    resourceVcdNsxvDhcpRelay(),    // 2.6
+			"vcd_nsxv_ip_set":        resourceVcdIpSet(),            // 2.6
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
@@ -153,7 +154,8 @@ func Provider() terraform.ResourceProvider {
 			"vcd_nsxv_dnat":          datasourceVcdNsxvDnat(),         // 2.5
 			"vcd_nsxv_snat":          datasourceVcdNsxvSnat(),         // 2.5
 			"vcd_nsxv_firewall_rule": datasourceVcdNsxvFirewallRule(), // 2.5
-			"vcd_ipset":              datasourceVcdIpSet(),            // 2.6
+			"vcd_nsxv_dhcp_relay":    datasourceVcdNsxvDhcpRelay(),    // 2.6
+			"vcd_nsxv_ip_set":        datasourceVcdIpSet(),            // 2.6
 		},
 
 		ConfigureFunc: providerConfigure,