+++ vcd/resource_vcd_ipset_test.go
@@ -41,40 +41,40 @@ func TestAccVcdIpSet(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
 		Providers:    testAccProviders,
 		PreCheck:     func() { testAccPreCheck(t) },
-		CheckDestroy: testAccCheckVcdIpSetDestroy("vcd_ipset.test-ipset", params["IpSetName"].(string)),
+		CheckDestroy: testAccCheckVcdIpSetDestroy("vcd_nsxv_ip_set.test-ipset", params["IpSetName"].(string)),
 		Steps: []resource.TestStep{
 			resource.TestStep{
 				Config: configText,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestMatchResourceAttr("vcd_ipset.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "name", t.Name()),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "description", "test-ip-set-description"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "is_inheritance_allowed", "true"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.#", "2"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.2555711295", "192.168.1.1"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.2329977041", "192.168.2.1"),
+					resource.TestMatchResourceAttr("vcd_nsxv_ip_set.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "name", t.Name()),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "description", "test-ip-set-description"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "is_inheritance_allowed", "true"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.#", "2"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.2555711295", "192.168.1.1"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.2329977041", "192.168.2.1"),
 
 					// Validate that datasource has all the same fields
-					resourceFieldsEqual("vcd_ipset.test-ipset", "data.vcd_ipset.test-ipset", []string{}),
+					resourceFieldsEqual("vcd_nsxv_ip_set.test-ipset", "data.vcd_nsxv_ip_set.test-ipset", []string{}),
 				),
 			},
 			resource.TestStep{
 				Config: configText1,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestMatchResourceAttr("vcd_ipset.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "name", t.Name()+"-changed"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "is_inheritance_allowed", "true"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "description", "test-ip-set-changed-description"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.#", "2"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.1441693733", "10.10.10.1"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.2766637002", "11.11.11.1"),
+					resource.TestMatchResourceAttr("vcd_nsxv_ip_set.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "name", t.Name()+"-changed"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "is_inheritance_allowed", "true"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "description", "test-ip-set-changed-description"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.#", "2"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.1441693733", "10.10.10.1"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.2766637002", "11.11.11.1"),
 
 					// Validate that datasource has all the same fields
-					resourceFieldsEqual("vcd_ipset.test-ipset", "data.vcd_ipset.test-ipset", []string{}),
+					resourceFieldsEqual("vcd_nsxv_ip_set.test-ipset", "data.vcd_nsxv_ip_set.test-ipset", []string{}),
 				),
 			},
 			resource.TestStep{
-				ResourceName:      "vcd_ipset.imported",
+				ResourceName:      "vcd_nsxv_ip_set.imported",
 				ImportState:       true,
 				ImportStateVerify: true,
 				ImportStateIdFunc: importStateIdOrgVdcObject(testConfig, "TestAccVcdIpSet-changed"),
@@ -82,30 +82,30 @@ func TestAccVcdIpSet(t *testing.T) {
 			resource.TestStep{
 				Config: configText2,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestMatchResourceAttr("vcd_ipset.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "name", t.Name()+"-changed2"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "is_inheritance_allowed", "false"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "description", "test-ip-set-changed-description"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.#", "2"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.2083356021", "1.1.1.1/24"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.3684382799", "10.10.10.100-10.10.10.110"),
+					resource.TestMatchResourceAttr("vcd_nsxv_ip_set.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "name", t.Name()+"-changed2"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "is_inheritance_allowed", "false"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "description", "test-ip-set-changed-description"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.#", "2"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.2083356021", "1.1.1.1/24"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.3684382799", "10.10.10.100-10.10.10.110"),
 					// Validate that datasource has all the same fields
-					resourceFieldsEqual("vcd_ipset.test-ipset", "data.vcd_ipset.test-ipset", []string{}),
+					resourceFieldsEqual("vcd_nsxv_ip_set.test-ipset", "data.vcd_nsxv_ip_set.test-ipset", []string{}),
 				),
 			},
 			resource.TestStep{
 				Config: configText2,
-				Taint:  []string{"vcd_ipset.test-ipset"}, // Force provisioning from scratch instead of update
+				Taint:  []string{"vcd_nsxv_ip_set.test-ipset"}, // Force provisioning from scratch instead of update
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestMatchResourceAttr("vcd_ipset.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "name", t.Name()+"-changed2"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "is_inheritance_allowed", "false"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "description", "test-ip-set-changed-description"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.#", "2"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.2083356021", "1.1.1.1/24"),
-					resource.TestCheckResourceAttr("vcd_ipset.test-ipset", "ip_addresses.3684382799", "10.10.10.100-10.10.10.110"),
+					resource.TestMatchResourceAttr("vcd_nsxv_ip_set.test-ipset", "id", regexp.MustCompile(`.*ipset-\d*$`)),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "name", t.Name()+"-changed2"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "is_inheritance_allowed", "false"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "description", "test-ip-set-changed-description"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.#", "2"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.2083356021", "1.1.1.1/24"),
+					resource.TestCheckResourceAttr("vcd_nsxv_ip_set.test-ipset", "ip_addresses.3684382799", "10.10.10.100-10.10.10.110"),
 					// Validate that datasource has all the same fields
-					resourceFieldsEqual("vcd_ipset.test-ipset", "data.vcd_ipset.test-ipset", []string{}),
+					resourceFieldsEqual("vcd_nsxv_ip_set.test-ipset", "data.vcd_nsxv_ip_set.test-ipset", []string{}),
 				),
 			},
 		},
@@ -139,7 +139,7 @@ func testAccCheckVcdIpSetDestroy(resource, ipSetName string) resource.TestCheckF
 }
 
 const testAccVcdIpSet = `
-resource "vcd_ipset" "test-ipset" {
+resource "vcd_nsxv_ip_set" "test-ipset" {
   org          = "{{.Org}}"
   vdc          = "{{.Vdc}}"
 
@@ -148,16 +148,16 @@ resource "vcd_ipset" "test-ipset" {
   ip_addresses = ["192.168.1.1","192.168.2.1"]
 }
 
-data "vcd_ipset" "test-ipset" {
+data "vcd_nsxv_ip_set" "test-ipset" {
 	org          = "{{.Org}}"
 	vdc          = "{{.Vdc}}"
   
-	name         = vcd_ipset.test-ipset.name
+	name         = vcd_nsxv_ip_set.test-ipset.name
 }
 `
 
 const testAccVcdIpSetUpdate = `
-resource "vcd_ipset" "test-ipset" {
+resource "vcd_nsxv_ip_set" "test-ipset" {
   org          = "{{.Org}}"
   vdc          = "{{.Vdc}}"
 
@@ -166,16 +166,16 @@ resource "vcd_ipset" "test-ipset" {
   ip_addresses = ["10.10.10.1","11.11.11.1"]
 }
 
-data "vcd_ipset" "test-ipset" {
+data "vcd_nsxv_ip_set" "test-ipset" {
 	org          = "{{.Org}}"
 	vdc          = "{{.Vdc}}"
   
-	name         = vcd_ipset.test-ipset.name
+	name         = vcd_nsxv_ip_set.test-ipset.name
 }
 `
 
 const testAccVcdIpSetUpdate2 = `
-resource "vcd_ipset" "test-ipset" {
+resource "vcd_nsxv_ip_set" "test-ipset" {
   org          = "{{.Org}}"
   vdc          = "{{.Vdc}}"
 
@@ -185,10 +185,10 @@ resource "vcd_ipset" "test-ipset" {
   ip_addresses           = ["1.1.1.1/24","10.10.10.100-10.10.10.110"]
 }
 
-data "vcd_ipset" "test-ipset" {
+data "vcd_nsxv_ip_set" "test-ipset" {
 	org          = "{{.Org}}"
 	vdc          = "{{.Vdc}}"
   
-	name         = vcd_ipset.test-ipset.name
+	name         = vcd_nsxv_ip_set.test-ipset.name
 }
 `